medius-server/requirements.txt
aiomysql==0.0.21
arrow==1.1.1
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cookiecutter==1.7.3
cryptography==3.4.7
databases==0.4.3
decorator==5.0.9
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
fastapi-bearer-auth==0.1.3
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.6.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyJWT==2.1.0
pymongo==3.12.0
PyMySQL==0.9.3
python-dateutil==2.8.2
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
text-unidecode==1.3
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
zipp==3.5.0
aiofiles==0.7.0